requirements.txt
python-dotenv==1.0.0
langchain-community==0.0.34 
langchain
openai
langchain-openai
gradio
google-generativeai
langchain-groq